move detailview from atom to molecule
